add admin service
