+++ dispatch.yaml
@@ -1,0 +1,3 @@
+dispatch:
+  - url: "*/admin/*"
+    service: admin
