+++ main.go
@@ -15,30 +15,6 @@ import (
 
 var redisPool *redis.Pool
 
-func incrementHandler(w http.ResponseWriter, r *http.Request) {
-	ctx := r.Context()
-
-	ctx, span := startSpan(ctx, "redis/increment")
-	defer span.End()
-
-	conn := redisPool.Get()
-	defer func() {
-		if err := conn.Close(); err != nil {
-			fmt.Printf("warning: conn.Close() err=%+v\n", err)
-		}
-	}()
-
-	counter, err := redis.Int(conn.Do("INCR", "visits"))
-	if err != nil {
-		http.Error(w, "Error incrementing visitor counter", http.StatusInternalServerError)
-		return
-	}
-	_, err = fmt.Fprintf(w, "Visitor number: %d", counter)
-	if err != nil {
-		fmt.Printf("failed write to response. err=%v\n", err)
-	}
-}
-
 func main() {
 	project, err := gcpmetadata.GetProjectID()
 	if err != nil {
@@ -66,6 +42,7 @@ func main() {
 	}, maxConnections)
 
 	http.Handle("/increment", ochttp.WithRouteTag(func() http.Handler { return http.HandlerFunc(incrementHandler) }(), "/gp"))
+	http.Handle("/admin/hello", ochttp.WithRouteTag(func() http.Handler { return http.HandlerFunc(adminHandler) }(), "/gp"))
 
 	port := os.Getenv("PORT")
 	if port == "" {
